feat: Reorganize downloads manager
