+++ lib/main.dart
@@ -36,6 +36,7 @@ import 'package:bluecherry_client/providers/mobile_view_provider.dart';
 import 'package:bluecherry_client/providers/server_provider.dart';
 import 'package:bluecherry_client/providers/settings_provider.dart';
 import 'package:bluecherry_client/providers/update_provider.dart';
+import 'package:bluecherry_client/screens/downloads/close_dialog.dart';
 import 'package:bluecherry_client/utils/app_links/app_links.dart' as app_links;
 import 'package:bluecherry_client/utils/logging.dart' as logging;
 import 'package:bluecherry_client/utils/methods.dart';
@@ -44,7 +45,6 @@ import 'package:bluecherry_client/utils/theme.dart';
 import 'package:bluecherry_client/utils/video_player.dart';
 import 'package:bluecherry_client/utils/window.dart';
 import 'package:bluecherry_client/widgets/desktop_buttons.dart';
-import 'package:bluecherry_client/widgets/downloads_manager.dart';
 import 'package:bluecherry_client/widgets/events/events_screen.dart';
 import 'package:bluecherry_client/widgets/home.dart';
 import 'package:bluecherry_client/widgets/multi_window/single_camera_window.dart';
@@ -251,11 +251,7 @@ class _UnityAppState extends State<UnityApp>
     if (isPreventClose && mounted && context.mounted) {
       final downloadsManager = context.read<DownloadsManager>();
       if (downloadsManager.downloading.isNotEmpty) {
-        final result = await showDialog<bool>(
-          context: context,
-          barrierDismissible: false,
-          builder: (context) => const CloseDownloadsDialog(),
-        );
+        final result = await showCloseDownloadsDialog(context);
         if (result == null || !result) {
           return;
         }
