+++ lib/screens/downloads/downloads_manager.dart
@@ -23,13 +23,12 @@ import 'package:bluecherry_client/models/event.dart';
 import 'package:bluecherry_client/providers/downloads_provider.dart';
 import 'package:bluecherry_client/providers/home_provider.dart';
 import 'package:bluecherry_client/providers/settings_provider.dart';
+import 'package:bluecherry_client/screens/downloads/indicators.dart';
 import 'package:bluecherry_client/utils/extensions.dart';
 import 'package:bluecherry_client/utils/methods.dart';
 import 'package:bluecherry_client/utils/theme.dart';
-import 'package:bluecherry_client/utils/widgets/squared_icon_button.dart';
 import 'package:bluecherry_client/utils/window.dart';
 import 'package:bluecherry_client/widgets/misc.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -336,128 +335,6 @@ class _DownloadTileState extends State<DownloadTile> {
   }
 }
 
-class DownloadProgressIndicator extends StatelessWidget {
-  const DownloadProgressIndicator({
-    super.key,
-    required this.progress,
-    this.color,
-  });
-
-  final DownloadProgress progress;
-
-  /// The color of the indicator.
-  ///
-  /// If not provided, the primary color is used instead.
-  final Color? color;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    if (isCupertino) {
-      return CupertinoActivityIndicator.partiallyRevealed(
-        progress: progress,
-      );
-    }
-
-    return Stack(children: [
-      Padding(
-        padding: const EdgeInsetsDirectional.all(8.0),
-        child: CircularProgressIndicator(
-          value: progress,
-          strokeWidth: 2.0,
-          color: color,
-        ),
-      ),
-      Center(
-        child: Icon(
-          Icons.download,
-          size: 14.0,
-          color: color ?? theme.colorScheme.primary,
-        ),
-      ),
-    ]);
-  }
-}
-
-/// The download indicator for the given event
-///
-/// See also:
-///  * [EventsScreenDesktop], which uses this to display the download status
-class DownloadIndicator extends StatelessWidget {
-  final Event event;
-
-  /// Whether to highlight the indicator with a white color and an outline border
-  final bool highlight;
-
-  /// Whether the indicator is small
-  final bool small;
-
-  const DownloadIndicator({
-    super.key,
-    required this.event,
-    this.highlight = false,
-    this.small = false,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final loc = AppLocalizations.of(context);
-
-    return SizedBox(
-      height: 40.0,
-      width: 40.0,
-      child: () {
-        if (event.isAlarm) {
-          return Icon(
-            Icons.warning,
-            color: theme.extension<UnityColors>()!.warningColor,
-            size: small ? 18.0 : null,
-          );
-        }
-
-        final downloads = context.watch<DownloadsManager>();
-
-        if (downloads.isEventDownloaded(event.id)) {
-          return SquaredIconButton(
-            onPressed: () {
-              context.read<HomeProvider>().toDownloads(event.id, context);
-            },
-            tooltip: loc.seeInDownloads,
-            icon: Icon(
-              Icons.download_done,
-              size: small ? 18.0 : null,
-              color: theme.extension<UnityColors>()!.successColor,
-            ),
-          );
-        }
-
-        if (downloads.isEventDownloading(event.id)) {
-          return DownloadProgressIndicator(
-            progress: downloads.downloading[downloads.downloading.keys
-                .firstWhere((e) => e.id == event.id)]!,
-            color: highlight ? Colors.amber : null,
-          );
-        }
-
-        if (event.mediaURL != null) {
-          return SquaredIconButton(
-            tooltip: loc.download,
-            onPressed: () => downloads.download(event),
-            icon: Icon(
-              Icons.download,
-              size: small ? 18.0 : 22.0,
-              color: highlight ? Colors.white : null,
-              shadows: highlight ? outlinedText() : null,
-            ),
-          );
-        }
-      }(),
-    );
-  }
-}
-
 class NoDownloads extends StatelessWidget {
   const NoDownloads({super.key});
 
@@ -484,67 +361,6 @@ class NoDownloads extends StatelessWidget {
               ..onTap = () => home.setTab(UnityTab.eventsScreen, context),
           ),
           textAlign: TextAlign.center,
-        ),
-      ],
-    );
-  }
-}
-
-/// The dialog that appears when the user tries to close the app when there are
-/// downloads in progress.
-class CloseDownloadsDialog extends StatefulWidget {
-  const CloseDownloadsDialog({super.key});
-
-  @override
-  State<CloseDownloadsDialog> createState() => _CloseDownloadsDialogState();
-}
-
-class _CloseDownloadsDialogState extends State<CloseDownloadsDialog> {
-  bool _closeWhenDone = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final downloadsManager = context.watch<DownloadsManager>();
-    final loc = AppLocalizations.of(context);
-    final navigator = Navigator.of(context);
-
-    return AlertDialog(
-      title: Text(loc.nDownloadsProgress(downloadsManager.downloading.length)),
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          for (final entry in downloadsManager.downloading.entries)
-            ListTile(
-              dense: true,
-              contentPadding: EdgeInsets.zero,
-              trailing: SizedBox.fromSize(
-                size: const Size(40.0, 40.0),
-                child: DownloadProgressIndicator(progress: entry.value),
-              ),
-              title: Text(entry.key.deviceName),
-              subtitle: Text(entry.key.server.name),
-            )
-        ],
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => navigator.pop<bool>(false),
-          child: Text(loc.cancel),
-        ),
-        OutlinedButton(
-          onPressed: _closeWhenDone ? null : () => navigator.pop<bool>(true),
-          child: Text(loc.closeAnyway),
-        ),
-        FilledButton(
-          onPressed: _closeWhenDone
-              ? null
-              : () async {
-                  setState(() => _closeWhenDone = true);
-                  await downloadsManager.downloadsCompleter?.future;
-                  navigator.pop<bool>(true);
-                },
-          child: Text(loc.closeWhenDone),
         ),
       ],
     );
